+++ dec/CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this crate will be documented in this file.
 The format is based on [Keep a Changelog], and this crate adheres to [Semantic
 Versioning].
 
-## Unreleased
+## 0.3.2 - 2021-03-10
 
 * Add the `set_exponent` method to `Context::<Decimal64>` and `Context::<Decimal128>`.